pyproject.toml
[project]
name = "databrick-datavault2-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "databricks-sql-connector>=4.2.5",
    "python-dotenv>=1.2.2",
]
